+++ vscode-extension/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "devbox" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.1.1]
+
+- Fixed documentation
+- Added devbox install command
+- Added devbox update command
+- Added devbox search command
+
 ## [0.1.0]
 
 - Added reopen in devbox shell environment feature that allows projects with devbox.json